==> notebooks/02_pre.ipynb <==
import pandas as pd
import sys
import os

sys.path.append(os.path.abspath(".."))

import src.utils.eda as eda
import src.pipeline.pr_v1 as pr1
import src.pipeline.pr_v2 as pr2
import src.pipeline.fe_v1 as fe1
import src.pipeline.fe_v2 as fe2
import src.pipeline.fe_v2_1 as fe21
import src.pipeline.fe_v2_2 as fe22
import src.pipeline.fe_v3 as fe3
import src.pipeline.fe_v4 as fe4
import src.pipeline.fe_v5 as fe5
import src.pipeline.fe_v6 as fe6
import src.pipeline.fe_v7 as fe7
import src.pipeline.fe_v8 as fe8
import src.pipeline.fe_v9 as fe9

# 生データの読み込み
train_data = pd.read_csv("../input/train.csv")
test_data = pd.read_csv("../input/test.csv")

# 前処理データの読み込み
train_data1 = pd.read_parquet("../artifacts/prepro/train_data1.parquet")
test_data1 = pd.read_parquet("../artifacts/prepro/test_data1.parquet")
train_data2 = pd.read_parquet("../artifacts/prepro/train_data2.parquet")
test_data2 = pd.read_parquet("../artifacts/prepro/test_data2.parquet")

# tr_df1
tr_df, test_df = fe1.feature_engineering(train_data1, test_data1)

tr_df.to_parquet("../artifacts/features/base/tr_df1.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df1.parquet", index=False)

# tr_df2
tr_df, test_df = fe2.feature_engineering(train_data1, test_data1)

tr_df.to_parquet("../artifacts/features/base/tr_df2.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df2.parquet", index=False)

# tr_df2_1
tr_df, test_df = fe21.feature_engineering(train_data1, test_data1)

tr_df.to_parquet("../artifacts/features/base/tr_df2_1.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df2_1.parquet", index=False)

# tr_df2_2
tr_df, test_df = fe22.feature_engineering(train_data1, test_data1)

tr_df.to_parquet("../artifacts/features/base/tr_df2_2.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df2_2.parquet", index=False)

# tr_df3
tr_df, test_df = fe3.feature_engineering(train_data1, test_data1)

tr_df.to_parquet("../artifacts/features/base/tr_df3.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df3.parquet", index=False)

# tr_df3_1
tr_df, test_df = fe3.feature_engineering(train_data2, test_data2)

tr_df.to_parquet("../artifacts/features/base/tr_df3_1.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df3_1.parquet", index=False)

# tr_df4
tr_df, test_df = fe4.feature_engineering(train_data1, test_data1)

tr_df.to_parquet("../artifacts/features/base/tr_df4.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df4.parquet", index=False)

# tr_df4_1
tr_df, test_df = fe4.feature_engineering(train_data2, test_data2)

tr_df.to_parquet("../artifacts/features/base/tr_df4_1.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df4_1.parquet", index=False)

# tr_df5
tr_df, test_df = fe5.feature_engineering(train_data2, test_data2)

tr_df.to_parquet("../artifacts/features/base/tr_df5.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df5.parquet", index=False)

# tr_df6
tr_df, test_df = fe6.feature_engineering(train_data2, test_data2)

tr_df.to_parquet("../artifacts/features/base/tr_df6.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df6.parquet", index=False)

# tr_df7
tr_df, test_df = fe7.feature_engineering(train_data2, test_data2)

tr_df.to_parquet("../artifacts/features/base/tr_df7.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df7.parquet", index=False)

# tr_df8
tr_df, test_df = fe8.feature_engineering(train_data1, test_data1)

tr_df.to_parquet("../artifacts/features/base/tr_df8.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df8.parquet", index=False)

# tr_df9
tr_df, test_df = fe9.feature_engineering(train_data1, test_data1)

tr_df.to_parquet("../artifacts/features/base/tr_df9.parquet", index=False)
test_df.to_parquet("../artifacts/features/base/test_df9.parquet", index=False)
