==> notebooks/05_lgbm_tuning.ipynb <==
import sys
import os
from pathlib import Path
from dotenv import load_dotenv
import pandas as pd
import optuna

sys.path.append(os.path.abspath(".."))

import src.models.lgbm.lgbm_optuna_optimizer as op
import src.utils.telegram as te

# データの読み込み
env_path = Path.cwd().parent / ".env"
load_dotenv(dotenv_path=env_path)
url = os.environ.get("OPTUNA_STORAGE_URL")

tr_df1 = pd.read_parquet("../artifacts/features/base/tr_df1.parquet")

# tuning
params = {
    "max_depth": 5,
    "num_leaves": 523,
    "min_child_samples": 5056,
    "min_split_gain": 9.275294835907687e-05,
    "feature_fraction": 0.43593953046851464,
    "bagging_fraction": 0.8924361138693251,
    "bagging_freq": 10,
    "lambda_l1": 1.6934155410667961,
    "lambda_l2": 0.6637926838138378
}

objective = op.create_objective(
    tr_df1,
    early_stopping_rounds=500,
    n_jobs=25
)

study = op.run_optuna_search(
    objective,
    n_trials=10,
    n_jobs=1,
    study_name="lgbm_v1",
    storage=url,
    sampler=optuna.samplers.TPESampler(
        n_startup_trials=20, seed=42),
    initial_params=params
)
te.send_telegram_message("LGBM Training complete!")
