==> notebooks/08_cb_analysis.ipynb <==
import sys, os
import pandas as pd
import optuna
import importlib

sys.path.append(os.path.abspath(".."))

import src.models.cb.cb_cv_trainer as cv
import src.utils.optuna_visualizer as opv

# データの読み込み
storage = "postgresql://optuna_user:optuna_pass@localhost:5432/optuna_db"

tr_df4 = pd.read_csv("../artifacts/features/tr_df4.csv")
test_df4 = pd.read_csv("../artifacts/features/test_df4.csv")

# cvの性能実験
params = {
    'iterations': 4200,
    'depth': 9,
    'learning_rate': 0.012311151478845978, 
    'l2_leaf_reg': 0.01182407665426616, 
    'random_strength': 0.2351783662718401, 
    'bagging_temperature': 0.8517467698288006, 
    'border_count': 129, 
    'min_data_in_leaf': 75, 
    'grow_policy': 'Depthwise',
    'loss_function': 'MultiClass',
}

trainer = cv.CBCVTrainer(params=params)
trainer.fit_one_fold(tr_df4)
