==> notebooks/09_logreg_analysis.ipynb <==
import sys
import os
from pathlib import Path
from dotenv import load_dotenv
import cudf
import numpy as np
import importlib
import optuna

sys.path.append(os.path.abspath(".."))

import src.models.logreg.logreg_cv_trainer as cv
import src.models.logreg.logreg_optuna_optimizer as op
import src.utils.optuna_visualizer as opv

# データの読み込み
env_path = Path.cwd().parent / ".env"
load_dotenv(dotenv_path=env_path)
url = os.environ.get("OPTUNA_STORAGE_URL")

tr_df2 = cudf.read_csv("../artifacts/features/tr_df2.csv")
tr_df4 = cudf.read_csv("../artifacts/features/tr_df4.csv")
test_df4 = cudf.read_csv("../artifacts/features/test_df4.csv")
test_df2 = cudf.read_csv("../artifacts/features/test_df2.csv")

# logreg_v1の確認
importlib.reload(opv)
study_name = "logreg_v6"

study = opv.OptunaVisualizer(study_name, url)
# study.visualize_optimization()
study.print_trials_table()
study.summarize_studies()
