==> notebooks/18_debug.ipynb <==
import sys
import os
from pathlib import Path
from dotenv import load_dotenv
import importlib
import pandas as pd
import numpy as np

sys.path.append(os.path.abspath(".."))

import src.models.mlp.mlp_cv_trainer as cv
import src.utils.optuna_visualizer as opv
import src.utils.telegram as te

# データの読み込み
env_path = Path.cwd().parent / ".env"
load_dotenv(dotenv_path=env_path)
url = os.environ.get("OPTUNA_STORAGE_URL")

l1_tr_df1 = pd.read_parquet("../artifacts/features/l1/l1_tr_df1.parquet").astype("float32")
l1_test_df1 = pd.read_parquet("../artifacts/features/l1/l1_test_df1.parquet").astype("float32")

# oofとの差を検証
importlib.reload(cv)
params = {
    "batch_size": 224,
    "lr": 0.0033251860643844526,
    "dropout_rate": 0.09,
    "activation": "ReLU",
    "hidden_dim1": 320,
    "hidden_dim2": 224
}
trainer = cv.MLPCVTrainer(**params)

trainer.fit_one_fold(l1_tr_df1)
te.send_telegram_message("MLP Training complete!")

# oof, test_predsの作成
importlib.reload(cv)
params = {
    "batch_size": 512,
    "lr": 0.006912279003734971,
    "dropout_rate": 0.09,
    "activation": "ReLU",
    "hidden_dim1": 256,
    "hidden_dim2": 192
}
trainer = cv.MLPCVTrainer(**params)

oof, test_preds = trainer.fit(l1_tr_df1, l1_test_df1)

# oof, test_predsの作成
params = {
    "batch_size": 736,
    "lr": 0.005365450324352026,
    "dropout_rate": 0.1,
    "activation": "ReLU",
    "hidden_dim1": 256,
    "hidden_dim2": 128
}
trainer = cv.MLPCVTrainer(**params)

oof, test_preds = trainer.fit(l1_tr_df1, l1_test_df1)
np.save("../artifacts/preds/l1/oof_single_7.npy", oof)
np.save("../artifacts/preds/l1/test_single_7.npy", test_preds)

# oof, test_predsの作成
params = {
    "batch_size": 320,
    "lr": 0.007158401724195687,
    "dropout_rate": 0.1,
    "activation": "ReLU",
    "hidden_dim1": 256,
    "hidden_dim2": 128
}
trainer = cv.MLPCVTrainer(**params)

oof, test_preds = trainer.fit(l1_tr_df1, l1_test_df1)
np.save("../artifacts/preds/l1/oof_single_8.npy", oof)
np.save("../artifacts/preds/l1/test_single_8.npy", test_preds)

# oof, test_predsの作成
params = {
    "batch_size": 1088,
    "lr": 0.00836670348003045,
    "dropout_rate": 0.07,
    "activation": "ReLU",
    "hidden_dim1": 416,
    "hidden_dim2": 352
}
trainer = cv.MLPCVTrainer(**params)

oof, test_preds = trainer.fit(l1_tr_df2, l1_test_df2)
np.save("../artifacts/preds/l1/oof_single_24.npy", oof)
np.save("../artifacts/preds/l1/test_single_24.npy", test_preds)

# oof, test_predsの作成
params = {
    "batch_size": 896,
    "lr": 0.007677555887250604,
    "dropout_rate": 0.05,
    "activation": "ReLU",
    "hidden_dim1": 384,
    "hidden_dim2": 352
}
trainer = cv.MLPCVTrainer(**params)

oof, test_preds = trainer.fit(l1_tr_df2, l1_test_df2)
np.save("../artifacts/preds/l1/oof_single_25.npy", oof)
np.save("../artifacts/preds/l1/test_single_25.npy", test_preds)

# oof, test_predsの作成
params = {
    "batch_size": 992,
    "lr": 0.0025298207248849675,
    "dropout_rate": 0.0,
    "activation": "ReLU",
    "hidden_dim1": 448,
    "hidden_dim2": 384
}
trainer = cv.MLPCVTrainer(**params)

oof, test_preds = trainer.fit(l1_tr_df2, l1_test_df2)
np.save("../artifacts/preds/l1/oof_single_26.npy", oof)
np.save("../artifacts/preds/l1/test_single_26.npy", test_preds)

# oof, test_predsの作成
params = {
    "batch_size": 1152,
    "lr": 0.008396872216648935,
    "dropout_rate": 0.07,
    "activation": "ReLU",
    "hidden_dim1": 288,
    "hidden_dim2": 224
}
trainer = cv.MLPCVTrainer(**params)

oof, test_preds = trainer.fit(l1_tr_df2, l1_test_df2)
np.save("../artifacts/preds/l1/oof_single_27.npy", oof)
np.save("../artifacts/preds/l1/test_single_27.npy", test_preds)

import src.utils.telegram as te
te.send_telegram_message("Training complete!")

# full train
params = {
    "batch_size": 960,
    "lr": 0.004888160547569001,
    "dropout_rate": 0.11,
    "activation": "ReLU",
    "hidden_dim1": 480,
    "hidden_dim2": 448,
    "hidden_dim3": 448
}
trainer = cv.MLPCVTrainer(**params)

test_preds = trainer.fit(tr_df9, test_df9)
np.save("../artifacts/preds/l1/test_full_9.npy", test_preds)
