==> notebooks/19_ens_tuning.ipynb <==
import sys
import os
from pathlib import Path
from dotenv import load_dotenv
import numpy as np
import optuna

sys.path.append(os.path.abspath(".."))

import src.utils.weight_optimizer as wo

# データの読み込み
env_path = Path.cwd().parent / ".env"
load_dotenv(dotenv_path=env_path)
url = os.environ.get("OPTUNA_STORAGE_URL")

# mlp
oof_single_7 = np.load("../artifacts/preds/l1/oof_single_7.npy")
oof_single_27 = np.load("../artifacts/preds/l1/oof_single_27.npy")

# xgb
oof_single_9 = np.load("../artifacts/preds/l1/oof_single_9.npy")
oof_single_12 = np.load("../artifacts/preds/l1/oof_single_12.npy")
oof_single_15 = np.load("../artifacts/preds/l1/oof_single_15.npy")

# cb
oof_single_14 = np.load("../artifacts/preds/base/oof_single_14.npy")
oof_single_21 = np.load("../artifacts/preds/l1/oof_single_21.npy")
oof_single_3 = np.load("../artifacts/preds/l1/oof_single_3.npy")
oof_single_19 = np.load("../artifacts/preds/l1/oof_single_19.npy")

# 最適な重みの探索
list_v1 = [
    oof_single_7, oof_single_27,  # mlp
    oof_single_9, oof_single_12, oof_single_15,  # xgb
    oof_single_21, oof_single_3, oof_single_19  # cb
]
list_v2 = [
    oof_single_7, oof_single_27,  # mlp
    oof_single_12, oof_single_15  # xgb
]

list_v3 = [
    oof_single_7, oof_single_27,  # mlp
    oof_single_12, oof_single_15,  # xgb
    oof_single_14
]
objective = wo.create_objective(list_v2)

study = wo.run_optuna_search(
    objective,
    n_trials=300,
    n_jobs=1,
    study_name="ens_v2",
    storage=url,
    sampler=optuna.samplers.TPESampler(
        n_startup_trials=100,
        seed=42)
)
